# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fitness AI API Documentation",
        "description": "L5 Swagger OpenApi description for the Fitness Backend",
        "contact": {
            "email": "contact@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/api/v1/plans/generate-workout": {
            "post": {
                "tags": [
                    "AI Plans"
                ],
                "summary": "Generate a personalized AI workout plan",
                "description": "Sends user preferences to Gemini AI and receives a structured workout plan that is saved to the database.",
                "operationId": "2a4a8e0f25293ceacde33c8447714ec6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "goal": {
                                        "type": "string",
                                        "example": "hypertrophy"
                                    },
                                    "muscles": {
                                        "type": "string",
                                        "example": "chest, triceps, shoulders"
                                    },
                                    "level": {
                                        "type": "string",
                                        "example": "intermediate"
                                    },
                                    "days_per_week": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "workout_time_minutes": {
                                        "type": "integer",
                                        "example": 60
                                    },
                                    "limitations": {
                                        "type": "string",
                                        "example": "shoulder injury on right side"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "home"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Workout plan generated and saved successfully"
                    },
                    "422": {
                        "description": "Validation error or AI processing failure"
                    }
                }
            }
        },
...
        "/api/v1/plans/{id}": {
            "get": {
                "tags": [
                    "AI Plans"
                ],
                "summary": "Get a specific plan by ID",
                "operationId": "90ba177b12b1fa90d22453b1bc81d084",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plan details"
                    },
                    "404": {
                        "description": "Plan not found"
                    }
                }
            }
        },
        "/api/v1/plans/{id}/activate": {
            "patch": {
                "tags": [
                    "AI Plans"
                ],
                "summary": "Accept and activate a draft plan",
                "operationId": "deadbf5fb156b59e3ca18618800fd746",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plan activated"
                    },
                    "404": {
                        "description": "Plan not found"
                    }
                }
            }
        },
        "/api/v1/plans/{id}/archive": {
            "patch": {
                "tags": [
                    "AI Plans"
                ],
                "summary": "Archive a plan",
                "operationId": "3ad47ecfa31328d24c7ccf105dc97be1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plan archived"
                    },
                    "404": {
                        "description": "Plan not found"
                    }
                }
            }
        },
        "/api/v1/plans/{id}/duplicate": {
            "post": {
                "tags": [
                    "AI Plans"
                ],
                "summary": "Duplicate an existing plan as a new draft",
                "operationId": "017c214b7a50b66b107820bb96be754c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Plan duplicated"
                    },
                    "404": {
                        "description": "Plan not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}